requirements.txt
allure-pytest==2.15.0
allure-python-commons==2.15.0
astroid==3.3.11
attrs==25.3.0
bandit==1.8.6
black==25.9.0
cachetools==6.2.0
certifi==2025.8.3
chardet==5.2.0
charset-normalizer==3.4.3
click==8.3.0
cssselect==1.3.0
cssutils==2.11.1
dill==0.4.0
emails==0.6
flake8==7.3.0
h11==0.16.0
idna==3.10
iniconfig==2.1.0
isort==6.1.0
Jinja2==3.1.6
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
more-itertools==10.8.0
mypy_extensions==1.1.0
outcome==1.3.0.post0
packaging==25.0
pathspec==0.12.1
pillow==12.0.0
platformdirs==4.5.0
pluggy==1.6.0
premailer==3.10.0
pycodestyle==2.14.0
pyflakes==3.4.0
Pygments==2.19.2
pylint==3.3.9
PySocks==1.7.1
pytest==8.4.2
pytest-html==4.1.1
pytest-json-report==1.5.0
pytest-metadata==3.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytokens==0.1.10
PyYAML==6.0.3
requests==2.32.5
rich==14.2.0
schedule==1.2.2
selenium==4.35.0
six==1.17.0
slack_sdk==3.36.0
sniffio==1.3.1
sortedcontainers==2.4.0
stevedore==5.5.0
tomlkit==0.13.3
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.1
urllib3==2.5.0
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
